Implement local sequence number counting lock-free
With this commit we use CAS instead of locking to guard the local
sequence number counter. We do this to reduce lock contention when a
large number of clients indexes data in Elasticsearch.

This commit only targets `LocalCheckpointTracker#generateSeqNo()` and
the related methods but leaves
`LocalCheckpointTracker#markSeqNoAsCompleted()` more or less in its
current state because we have seen it to cause less lock contention than
the other method and it is also harder to implement lock-free.

Closes #26339